separating groups from users
