+++ services/web/server/src/simcore_service_webserver/security_roles.py
@@ -61,7 +61,7 @@ ROLES_PERMISSIONS = {
             # "preferences.apikey.delete"
             "user.tokens.*",  # "preferences.token.create",
             # "preferences.token.delete"
-            "user.groups.*",
+            "groups.*",
             "tag.crud.*"  # "preferences.tag"
             # NOTE: All services* are not necessary since it only requires login
             # and there is no distinction among logged in users.
