+++ services/web/server/src/simcore_service_webserver/users_exceptions.py
@@ -1,5 +1,7 @@
 """Defines the different exceptions that may arise in the projects subpackage"""
 
+from typing import Optional
+
 
 class UsersException(Exception):
     """Basic exception for errors raised in projects"""
@@ -11,33 +13,12 @@ class UsersException(Exception):
 class UserNotFoundError(UsersException):
     """User in group was not found in DB"""
 
-    def __init__(self, uid: int):
-        super().__init__(f"User id {uid} not found")
+    def __init__(self, *, uid: Optional[int] = None, email: Optional[str] = None):
+        super().__init__(
+            f"User id {uid} not found" if uid else f"User with email {email} not found"
+        )
         self.uid = uid
-
-
-class GroupNotFoundError(UsersException):
-    """Group was not found in DB"""
-
-    def __init__(self, gid: int):
-        super().__init__(f"Group with id {gid} not found")
-        self.gid = gid
-
-
-class UserInGroupNotFoundError(UsersException):
-    """User in group was not found in DB"""
-
-    def __init__(self, gid: int, uid: int):
-        super().__init__(f"User id {uid} in Group {gid} not found")
-        self.gid = gid
-        self.uid = uid
-
-
-class UserInsufficientRightsError(UsersException):
-    """User has not sufficient rights"""
-
-    def __init__(self, msg: str):
-        super().__init__(msg)
+        self.email = email
 
 
 class TokenNotFoundError(UsersException):
