+++ services/web/server/tests/unit/with_dbs/test_users.py
@@ -24,16 +24,16 @@ from pytest_simcore.helpers.utils_tokens import (
 )
 from servicelib.application import create_safe_application
 from simcore_service_webserver.db import APP_DB_ENGINE_KEY, setup_db
+from simcore_service_webserver.groups_api import (
+    DEFAULT_GROUP_OWNER_ACCESS_RIGHTS,
+    DEFAULT_GROUP_READ_ACCESS_RIGHTS,
+)
 from simcore_service_webserver.login import setup_login
 from simcore_service_webserver.rest import setup_rest
 from simcore_service_webserver.security import setup_security
 from simcore_service_webserver.security_roles import UserRole
 from simcore_service_webserver.session import setup_session
 from simcore_service_webserver.users import setup_users
-from simcore_service_webserver.users_api import (
-    DEFAULT_GROUP_OWNER_ACCESS_RIGHTS,
-    DEFAULT_GROUP_READ_ACCESS_RIGHTS,
-)
 
 ## BUG FIXES #######################################################
 from simcore_service_webserver.utils import gravatar_hash
